Build/Test Tools: Raise timeout value when running tests on PHP 8.4.
The test suite when run on PHP 8.4 with MySQL is currently taking 2-3x the amount of time to run. The jobs are regularly hitting the conservative `20` minute time out configured to prevent runaway jobs.

While this performance regression is investigated, this increases the timeout value to `30` to avoid running into unnecessary failures now that the issue has been discovered.

Props johnbillion.
See #63026.

git-svn-id: https://develop.svn.wordpress.org/trunk@59873 602fd350-edb4-49c9-b593-d223f7449a82