Media: Allow super-admin to access files of archived sites.
This changeset fixes an issue in multisite installations where archived sites remain accessible to network administrators, but the associated files do not. The previous implementation was checking if the blog is archived, marked as spam, or deleted, to subsequently return a 404 error for file requests. However, this did not account for network administrators who should retain access to these files.

Props antwortzeit, jeremyfelt, debarghyabanerjee, audrasjb.
Fixes #36803.



git-svn-id: https://develop.svn.wordpress.org/trunk@59967 602fd350-edb4-49c9-b593-d223f7449a82