Coding Standards: Remove unnecessary conditional in `WP_List_Table::print_column_headers()`.
This commit removes an unnecessary `empty()` check, where the `$class` variable is primed at the top of the `foreach` loop and will always be a non-empty array.

To maintain readability of the source of the `$class` attribute without the unnecessary `empty()` check, the resulting variable has been renamed to `$class_attr`, instead of reusing the array.

Follow-up to [8923], [15491].

Props justlevine.
See #52217.

git-svn-id: https://develop.svn.wordpress.org/trunk@60132 602fd350-edb4-49c9-b593-d223f7449a82