Build/Test Tools: Monitor default themes for dependency updates.
There are currently 3 default themes with build processes: Twenty Twenty-One, Twenty Twenty, and Twenty Nineteen. This adds a configuration for each to the Dependabot configuration to ensure pull requests are opened when updates become available.

Because they have a low number of dependencies, they will all be updated in the same pull request to start. Groups can be configured later if this presents problems.

See #63170, #63170.

git-svn-id: https://develop.svn.wordpress.org/trunk@60347 602fd350-edb4-49c9-b593-d223f7449a82