Twenty Twenty-One: Add nav menu toggle behaviors at `DOMContentLoaded` instead of at the `load` event.
The `load` event (on `window`) fires after all resources on a page have been loaded, including images. This means that 
on a slow connection, the mobile nav menu button may be rendered yet unresponsive to taps between when it first appears 
and when the page has fully loaded. This is rectified by switching to the `DOMContentLoaded` event so that the nav menu
behaviors are attached as soon as the DOM has fully loaded.

Props gernberg, westonruter, sabernhardt.
Fixes #63613.


git-svn-id: https://develop.svn.wordpress.org/trunk@60352 602fd350-edb4-49c9-b593-d223f7449a82