Canonical: prevent possible "Undefined array key" PHP error in `redirect_canonical()`, if the `path` or `query` array keys have gone missing.
This commit checks for the existence of (and re-adds if necessary) the `path` and `query` array keys after the `$redirect` variable has potentially been replaced by a second call to `parse_url()`.

This may happen when redirecting back to the root domain, without any path ('/example/`) or query (`?example=1`) to parse.

Props chrismattix, dhruvang21.

Fixes #63733.

git-svn-id: https://develop.svn.wordpress.org/trunk@60496 602fd350-edb4-49c9-b593-d223f7449a82