Administration: Reduce mobile item spacing in `.misc-pub-section`.
On small viewports, the total vertical padding of `.misc-pub-section` was adjusted from 14px to 40px.

This significant different made for difficult reading of information in the Post Publish metabox, the Media details meta box, and the comment editing metabox.

Shift the total vertical padding to 24px to improve the readability of metaboxes on mobile.

Props mikinc860, dhrumilk, debarghyabanerjee, mukesh27, hbhalodia, sabernhardt, joedolson.
Fixes #63571.

git-svn-id: https://develop.svn.wordpress.org/trunk@60619 602fd350-edb4-49c9-b593-d223f7449a82