Coding Standards: Move `count()` usage in `wp_dashboard_recent_comments()`.
While the rule to discourage using functions like `count()` in a loop condition is a recommendation/best practice rule from the `WordPress-Extra` ruleset and does not directly apply to WordPress core, this is intended as a minor readability and code clarity improvement.

Follow-up to [10090], [17556], [20609], [26144].

Props krunal265, johnbillion, audrasjb, dhruvang21, SergeyBiryukov.
Fixes #56499.

git-svn-id: https://develop.svn.wordpress.org/trunk@60643 602fd350-edb4-49c9-b593-d223f7449a82